vm: lib -> src rename code adjustments, added src folder to distribution build (package.json)
