+++ packages/vm/.nycrc
@@ -1,6 +1,6 @@
 {
   "extends": "../../config/nyc.json",
   "include": [
-    "lib/**/*.ts"
+    "src/**/*.ts"
   ]
 }
