+++ packages/vm/package.json
@@ -10,7 +10,8 @@
   ],
   "files": [
     "dist",
-    "dist.browser"
+    "dist.browser",
+    "src"
   ],
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
