+++ packages/vm/tests/api/utils.ts
@@ -1,7 +1,7 @@
 import { Account, Address, BN } from 'ethereumjs-util'
 import Blockchain from '@ethereumjs/blockchain'
-import VM from '../../lib/index'
-import { VMOpts } from '../../lib'
+import VM from '../../src/index'
+import { VMOpts } from '../../src'
 import { Block } from '@ethereumjs/block'
 import { TransactionFactory } from '@ethereumjs/tx'
 import Common from '@ethereumjs/common'
